requirements.txt
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
joblib>=1.3.0
imbalanced-learn>=0.11.0
mygene>=3.2.2
requests>=2.28.0
